third_party/WebKit/LayoutTests/fast/dom/NodeList/nodelist-item-with-name-expected.txt

This tests that items in a NodeList can be retrieved directly by name.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

text1
text2


PASS div1 instanceof HTMLDivElement is true
PASS div2 instanceof HTMLDivElement is true
PASS div1.innerHTML is 'text1'
PASS div2.innerHTML is 'text2'
PASS successfullyParsed is true

TEST COMPLETE
